Moved Pod and container builder funcs to dedicated files
