+++ pkg/builder/statefulset_builder_test.go
@@ -287,498 +287,3 @@ func TestMaxScaleStatefulSetMeta(t *testing.T) {
 		})
 	}
 }
-
-func TestMariadbPodMeta(t *testing.T) {
-	builder := newTestBuilder()
-	objMeta := metav1.ObjectMeta{
-		Name: "mariadb-obj",
-	}
-	tests := []struct {
-		name     string
-		mariadb  *mariadbv1alpha1.MariaDB
-		opts     []mariadbOpt
-		wantMeta *mariadbv1alpha1.Metadata
-	}{
-		{
-			name: "empty",
-			mariadb: &mariadbv1alpha1.MariaDB{
-				ObjectMeta: objMeta,
-			},
-			opts: nil,
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "mariadb",
-					"app.kubernetes.io/instance": "mariadb-obj",
-				},
-				Annotations: map[string]string{},
-			},
-		},
-		{
-			name: "inherit meta",
-			mariadb: &mariadbv1alpha1.MariaDB{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MariaDBSpec{
-					InheritMetadata: &mariadbv1alpha1.Metadata{
-						Labels: map[string]string{
-							"sidecar.istio.io/inject": "false",
-						},
-						Annotations: map[string]string{
-							"database.myorg.io": "mariadb",
-						},
-					},
-				},
-			},
-			opts: nil,
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "mariadb",
-					"app.kubernetes.io/instance": "mariadb-obj",
-					"sidecar.istio.io/inject":    "false",
-				},
-				Annotations: map[string]string{
-					"database.myorg.io": "mariadb",
-				},
-			},
-		},
-		{
-			name: "HA",
-			mariadb: &mariadbv1alpha1.MariaDB{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MariaDBSpec{
-					Galera: &mariadbv1alpha1.Galera{
-						Enabled: true,
-					},
-				},
-			},
-			opts: nil,
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "mariadb",
-					"app.kubernetes.io/instance": "mariadb-obj",
-				},
-				Annotations: map[string]string{
-					"k8s.mariadb.com/mariadb": "mariadb-obj",
-					"k8s.mariadb.com/galera":  "",
-				},
-			},
-		},
-		{
-			name: "Pod meta",
-			mariadb: &mariadbv1alpha1.MariaDB{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MariaDBSpec{
-					PodTemplate: mariadbv1alpha1.PodTemplate{
-						PodMetadata: &mariadbv1alpha1.Metadata{
-							Labels: map[string]string{
-								"sidecar.istio.io/inject": "false",
-							},
-							Annotations: map[string]string{
-								"database.myorg.io": "mariadb",
-							},
-						},
-					},
-				},
-			},
-			opts: nil,
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "mariadb",
-					"app.kubernetes.io/instance": "mariadb-obj",
-					"sidecar.istio.io/inject":    "false",
-				},
-				Annotations: map[string]string{
-					"database.myorg.io": "mariadb",
-				},
-			},
-		},
-		{
-			name: "extra meta",
-			mariadb: &mariadbv1alpha1.MariaDB{
-				ObjectMeta: objMeta,
-			},
-			opts: []mariadbOpt{
-				withMeta(&mariadbv1alpha1.Metadata{
-					Labels: map[string]string{
-						"sidecar.istio.io/inject": "false",
-					},
-					Annotations: map[string]string{
-						"database.myorg.io": "mariadb",
-					},
-				}),
-			},
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "mariadb",
-					"app.kubernetes.io/instance": "mariadb-obj",
-					"sidecar.istio.io/inject":    "false",
-				},
-				Annotations: map[string]string{
-					"database.myorg.io": "mariadb",
-				},
-			},
-		},
-		{
-			name: "inherit and Pod meta",
-			mariadb: &mariadbv1alpha1.MariaDB{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MariaDBSpec{
-					InheritMetadata: &mariadbv1alpha1.Metadata{
-						Annotations: map[string]string{
-							"database.myorg.io": "mariadb",
-						},
-					},
-					PodTemplate: mariadbv1alpha1.PodTemplate{
-						PodMetadata: &mariadbv1alpha1.Metadata{
-							Labels: map[string]string{
-								"sidecar.istio.io/inject": "false",
-							},
-						},
-					},
-				},
-			},
-			opts: nil,
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "mariadb",
-					"app.kubernetes.io/instance": "mariadb-obj",
-					"sidecar.istio.io/inject":    "false",
-				},
-				Annotations: map[string]string{
-					"database.myorg.io": "mariadb",
-				},
-			},
-		},
-		{
-			name: "extra override Pod meta",
-			mariadb: &mariadbv1alpha1.MariaDB{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MariaDBSpec{
-					PodTemplate: mariadbv1alpha1.PodTemplate{
-						PodMetadata: &mariadbv1alpha1.Metadata{
-							Labels: map[string]string{
-								"sidecar.istio.io/inject": "false",
-							},
-						},
-					},
-				},
-			},
-			opts: []mariadbOpt{
-				withMeta(&mariadbv1alpha1.Metadata{
-					Labels: map[string]string{
-						"sidecar.istio.io/inject": "true",
-					},
-				}),
-			},
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "mariadb",
-					"app.kubernetes.io/instance": "mariadb-obj",
-					"sidecar.istio.io/inject":    "true",
-				},
-				Annotations: map[string]string{},
-			},
-		},
-		{
-			name: "Pod override inherit meta",
-			mariadb: &mariadbv1alpha1.MariaDB{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MariaDBSpec{
-					PodTemplate: mariadbv1alpha1.PodTemplate{
-						PodMetadata: &mariadbv1alpha1.Metadata{
-							Labels: map[string]string{
-								"sidecar.istio.io/inject": "false",
-							},
-						},
-					},
-					InheritMetadata: &mariadbv1alpha1.Metadata{
-						Labels: map[string]string{
-							"sidecar.istio.io/inject": "true",
-						},
-						Annotations: map[string]string{
-							"database.myorg.io": "mariadb",
-						},
-					},
-				},
-			},
-			opts: nil,
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "mariadb",
-					"app.kubernetes.io/instance": "mariadb-obj",
-					"sidecar.istio.io/inject":    "false",
-				},
-				Annotations: map[string]string{
-					"database.myorg.io": "mariadb",
-				},
-			},
-		},
-		{
-			name: "without selector labels",
-			mariadb: &mariadbv1alpha1.MariaDB{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MariaDBSpec{
-					PodTemplate: mariadbv1alpha1.PodTemplate{
-						PodMetadata: &mariadbv1alpha1.Metadata{
-							Labels: map[string]string{
-								"sidecar.istio.io/inject": "false",
-							},
-						},
-					},
-				},
-			},
-			opts: []mariadbOpt{
-				withMariadbSelectorLabels(false),
-			},
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"sidecar.istio.io/inject": "false",
-				},
-				Annotations: map[string]string{},
-			},
-		},
-		{
-			name: "all",
-			mariadb: &mariadbv1alpha1.MariaDB{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MariaDBSpec{
-					InheritMetadata: &mariadbv1alpha1.Metadata{
-						Annotations: map[string]string{
-							"database.myorg.io": "mariadb",
-						},
-					},
-					Galera: &mariadbv1alpha1.Galera{
-						Enabled: true,
-					},
-					PodTemplate: mariadbv1alpha1.PodTemplate{
-						PodMetadata: &mariadbv1alpha1.Metadata{
-							Labels: map[string]string{
-								"sidecar.istio.io/inject": "false",
-							},
-						},
-					},
-				},
-			},
-			opts: []mariadbOpt{
-				withMeta(&mariadbv1alpha1.Metadata{
-					Annotations: map[string]string{
-						"sidecar.istio.io/inject": "false",
-					},
-				}),
-			},
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "mariadb",
-					"app.kubernetes.io/instance": "mariadb-obj",
-					"sidecar.istio.io/inject":    "false",
-				},
-				Annotations: map[string]string{
-					"database.myorg.io":       "mariadb",
-					"k8s.mariadb.com/mariadb": "mariadb-obj",
-					"k8s.mariadb.com/galera":  "",
-					"sidecar.istio.io/inject": "false",
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			podTpl, err := builder.mariadbPodTemplate(tt.mariadb, tt.opts...)
-			if err != nil {
-				t.Fatalf("unexpected error building MariaDB Pod template: %v", err)
-			}
-			assertMeta(t, &podTpl.ObjectMeta, tt.wantMeta.Labels, tt.wantMeta.Annotations)
-		})
-	}
-}
-
-func TestMaxScalePodTemplate(t *testing.T) {
-	builder := newTestBuilder()
-	objMeta := metav1.ObjectMeta{
-		Name: "maxscale-obj",
-	}
-	tests := []struct {
-		name     string
-		maxscale *mariadbv1alpha1.MaxScale
-		wantMeta *mariadbv1alpha1.Metadata
-	}{
-		{
-			name: "empty",
-			maxscale: &mariadbv1alpha1.MaxScale{
-				ObjectMeta: objMeta,
-			},
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "maxscale",
-					"app.kubernetes.io/instance": "maxscale-obj",
-				},
-				Annotations: map[string]string{},
-			},
-		},
-		{
-			name: "inherit meta",
-			maxscale: &mariadbv1alpha1.MaxScale{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MaxScaleSpec{
-					InheritMetadata: &mariadbv1alpha1.Metadata{
-						Labels: map[string]string{
-							"sidecar.istio.io/inject": "false",
-						},
-						Annotations: map[string]string{
-							"database.myorg.io": "mariadb",
-						},
-					},
-				},
-			},
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "maxscale",
-					"app.kubernetes.io/instance": "maxscale-obj",
-					"sidecar.istio.io/inject":    "false",
-				},
-				Annotations: map[string]string{
-					"database.myorg.io": "mariadb",
-				},
-			},
-		},
-		{
-			name: "Pod meta",
-			maxscale: &mariadbv1alpha1.MaxScale{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MaxScaleSpec{
-					PodTemplate: mariadbv1alpha1.PodTemplate{
-						PodMetadata: &mariadbv1alpha1.Metadata{
-							Labels: map[string]string{
-								"sidecar.istio.io/inject": "false",
-							},
-							Annotations: map[string]string{
-								"database.myorg.io": "mariadb",
-							},
-						},
-					},
-				},
-			},
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "maxscale",
-					"app.kubernetes.io/instance": "maxscale-obj",
-					"sidecar.istio.io/inject":    "false",
-				},
-				Annotations: map[string]string{
-					"database.myorg.io": "mariadb",
-				},
-			},
-		},
-		{
-			name: "inherit and Pod meta",
-			maxscale: &mariadbv1alpha1.MaxScale{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MaxScaleSpec{
-					InheritMetadata: &mariadbv1alpha1.Metadata{
-						Annotations: map[string]string{
-							"database.myorg.io": "mariadb",
-						},
-					},
-					PodTemplate: mariadbv1alpha1.PodTemplate{
-						PodMetadata: &mariadbv1alpha1.Metadata{
-							Labels: map[string]string{
-								"sidecar.istio.io/inject": "false",
-							},
-						},
-					},
-				},
-			},
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "maxscale",
-					"app.kubernetes.io/instance": "maxscale-obj",
-					"sidecar.istio.io/inject":    "false",
-				},
-				Annotations: map[string]string{
-					"database.myorg.io": "mariadb",
-				},
-			},
-		},
-		{
-			name: "Pod override inherit meta",
-			maxscale: &mariadbv1alpha1.MaxScale{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MaxScaleSpec{
-					InheritMetadata: &mariadbv1alpha1.Metadata{
-						Labels: map[string]string{
-							"sidecar.istio.io/inject": "true",
-						},
-						Annotations: map[string]string{
-							"database.myorg.io": "mariadb",
-						},
-					},
-					PodTemplate: mariadbv1alpha1.PodTemplate{
-						PodMetadata: &mariadbv1alpha1.Metadata{
-							Labels: map[string]string{
-								"sidecar.istio.io/inject": "false",
-							},
-						},
-					},
-				},
-			},
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "maxscale",
-					"app.kubernetes.io/instance": "maxscale-obj",
-					"sidecar.istio.io/inject":    "false",
-				},
-				Annotations: map[string]string{
-					"database.myorg.io": "mariadb",
-				},
-			},
-		},
-		{
-			name: "all",
-			maxscale: &mariadbv1alpha1.MaxScale{
-				ObjectMeta: objMeta,
-				Spec: mariadbv1alpha1.MaxScaleSpec{
-					InheritMetadata: &mariadbv1alpha1.Metadata{
-						Labels: map[string]string{
-							"k8s.mariadb.com": "test",
-						},
-						Annotations: map[string]string{
-							"k8s.mariadb.com": "test",
-						},
-					},
-					PodTemplate: mariadbv1alpha1.PodTemplate{
-						PodMetadata: &mariadbv1alpha1.Metadata{
-							Labels: map[string]string{
-								"sidecar.istio.io/inject": "false",
-							},
-							Annotations: map[string]string{
-								"database.myorg.io": "mariadb",
-							},
-						},
-					},
-				},
-			},
-			wantMeta: &mariadbv1alpha1.Metadata{
-				Labels: map[string]string{
-					"app.kubernetes.io/name":     "maxscale",
-					"app.kubernetes.io/instance": "maxscale-obj",
-					"sidecar.istio.io/inject":    "false",
-					"k8s.mariadb.com":            "test",
-				},
-				Annotations: map[string]string{
-					"k8s.mariadb.com":   "test",
-					"database.myorg.io": "mariadb",
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			podTpl, err := builder.maxscalePodTemplate(tt.maxscale)
-			if err != nil {
-				t.Fatalf("unexpected error building MaxScale Pod template: %v", err)
-			}
-			assertMeta(t, &podTpl.ObjectMeta, tt.wantMeta.Labels, tt.wantMeta.Annotations)
-		})
-	}
-}
